Fix remaining clang warnings

+++ src/printer.h
@@ -8,7 +8,7 @@ typedef struct printer_params_t {
   queue_t *watchdog_queue;
   queue_t *logger_queue;
   FILE *output_file;
-  int nprocs;
+  size_t nprocs;
 } printer_params_t;
 
 void *printer_thread(void *printer_params);
